Use NSInteger, NSUInteger in preparation for building 64-bit.
A few places were switched to explicitly sized types instead - int32_t, uint32_t, etc.

+++ Frameworks/GnuSpeech/Archiving/MXMLIgnoreTreeDelegate.h
@@ -7,7 +7,7 @@
 
 @interface MXMLIgnoreTreeDelegate : NSObject
 {
-    int depth;
+    NSUInteger depth;
 }
 
 - (id)init;
